Seperate out Atttribute and Variant in different component || Remove added attribute || For adding variant fix issue for adding attribute || Add product API call and added loader
