+++ src/store/slice/productSlice.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { PRODUCTS_API_URL } from "../../utils/endpoints";
 import axios from '../../config/axios';
+import { toast } from "react-toastify";
 
 // Async actions
 export const fetchProducts = createAsyncThunk("get/fetchProducts", async (params) => {
@@ -10,12 +11,9 @@ export const fetchProducts = createAsyncThunk("get/fetchProducts", async (params
   }
 );
 
-export const addProducts = createAsyncThunk("add/addProducts", async (body) => {
+export const addProducts = createAsyncThunk("add/addProducts", async ({body, resetForm}) => {  
     // Create product
-    const response = await axios.post(PRODUCTS_API_URL, { ...body }, { headers: {
-      'Content-Type': 'application/json;charset=UTF-8',
-      "Access-Control-Allow-Origin": "*",
-  }});
+    const response = await axios.post(PRODUCTS_API_URL, { ...body });
     return response.data;
   }
 );
@@ -33,6 +31,10 @@ const initialState = {
     isError: false,
     data: {},
   },
+  createAndUpdateProduct: {
+    isLoading: false,
+    isError: false,
+  },
 };
 
 const productSlice = createSlice({
@@ -42,6 +44,7 @@ const productSlice = createSlice({
     // normal reducer functions go here
   },
   extraReducers: (builder) => {
+    // Case for product listing
     builder.addCase(fetchProducts.pending, (state, action) => {
         state.productsList.isLoading = true;
     });
@@ -51,6 +54,21 @@ const productSlice = createSlice({
     });
     builder.addCase(fetchProducts.rejected, (state, action) => {
         state.productsList.isLoading = false;
+        state.productsList.isError = true;
+    });
+
+  // Case for Create/Update product 
+    builder.addCase(addProducts.pending, (state, action) => {      
+        state.createAndUpdateProduct.isLoading = true;
+    });
+    builder.addCase(addProducts.fulfilled, (state, {meta}) => {    
+      state.createAndUpdateProduct.isLoading = false;
+      meta.arg.resetForm();
+      toast.success("Product created Successfully!");
+    });
+    builder.addCase(addProducts.rejected, (state, action) => {
+        state.createAndUpdateProduct.isLoading = false;
+        state.createAndUpdateProduct.isError = true;
     });
   },
 });
