+++ src/utils/constant.js
@@ -8,3 +8,15 @@ export const STRING_TYPE = 'STRING';
 export const LINK_TYPE = 'LINK';
 export const ENUM_TYPE = 'ENUM';
 export const BOOLEAN_TYPE = 'BOOLEAN';
+
+
+export const requiredKeyForCreateProduct = {  
+  "listingImages": ["5674788612","56747886154"],
+  "storeId": "0f3e43ce-0a1b-4a15-a85d-aad6b5fc780f",
+  "thumbnail": "456456zmcjbbhs",
+  "video": "546546fdfdfdb",
+  "returnExchangePolicyId": "Return and Exchange Policy",
+}
+
+export const SAVE = 'Save';
+export const PROCESSING = 'Processing...';
